refactor model selector filtering and tighten matcher
